requirements.txt
yfinance>=1.3.0
curl_cffi>=0.7.0
pandas==2.2.2
python-dotenv==1.0.1
pytz==2024.1
streamlit==1.41.1
plotly==5.24.1
